feat;djangorestframework&modify_board

+++ quest/grade/views.py
@@ -1,9 +1,18 @@
 from django.contrib.auth import authenticate, login
 from django.contrib import messages
 from django.shortcuts import get_object_or_404, redirect, render
+from grade.forms import BoardForm, GradeForm, ReportForm, ReportProblemForm,ProfileForm,CustomUserCreationForm
+from grade.models import Board, Grade, Report, ReportProblem,Profile,CustomUser
+from django.contrib.auth.decorators import login_required
+from django.db.models import Avg
+import django_filters
+from rest_framework import viewsets, filters
+from .models import Board
+from .serializers import BoardSerializer
 from .forms import BoardForm, ProfileForm, CustomUserCreationForm
 from .models import Board, Grade, Report, ReportProblem, Profile
 from django.contrib.auth.decorators import login_required
+
 
 
 @login_required
@@ -84,9 +93,11 @@ def new_board(request):
                   {"form": form, "boards": boards})
 
 
-def delete_board(request):
-    board = Board.objects.filter(user=request.user)
-    board.delete()
+def delete_board(request, board_id):
+    board = get_object_or_404(Board, id=board_id)
+    if board.user == request.user:
+        board.delete()
+
     return redirect("grade:new_board")
 
 
@@ -117,5 +128,10 @@ def update_profile(request):
             return redirect("grade:index")
     else:
         form = ProfileForm(instance=profileUser)
-    return render(request, "grade/update_profile.html",
-                  {"form": form, "photos": photos})
+    return render(request, "grade/update_profile.html", {"form": form, "photos": photos})
+
+
+class BoardViewSet(viewsets.ModelViewSet):
+    queryset = Board.objects.all()
+    serializer_class = BoardSerializer
+
